requirements1.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
channels==4.0.0
chardet==3.0.4
Django==4.1.5
django-debug-toolbar==2.2
django-model-utils==4.3.1
django-private-chat2==1.0.2
Faker==12.0.1
idna==2.8
isort==5.11.4
mixer==7.1.2
more-itertools==9.0.0
packaging==22.0
Pillow==9.3.0
pluggy==0.13.1
py==1.11.0
pytest==5.3.5
pytest-django==3.8.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
pytz==2022.7
requests==2.22.0
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.4.3
urllib3==1.25.11
wcwidth==0.2.5
